Removed uneccesary sysouts, and added documentation.
When pressed start, setUpGame() establishes a connection with the other players. When a new player is connected it is given playernumber and how many players in the game. The game becoming a server waits for all players to connect (given in terminal), and then send the deck. The players let the server know when it has chosen cards, and when all chosen cards are sent to each player the server sends message to start the turn.